requirements.txt
asgiref==3.5.2
Django==3.2.16
django-filter==22.1
djangorestframework==3.14.0
environs==9.5.0
marshmallow==3.19.0
packaging==21.3
Pillow==9.3.0
pkg_resources==0.0.0
pyparsing==3.0.9
python-dotenv==0.21.0
pytz==2022.6
sqlparse==0.4.3
typing_extensions==4.4.0
